Scrit de download of spark 3.0.1 and datasets, runner configure spark and lauch jupyter lab

+++ BDA/download_data.py
@@ -1,4 +1,5 @@
 from pathlib import Path, PurePath
+import subprocess
 import configparser
 import os
 import requests
@@ -13,7 +14,7 @@ logger = logging.getLogger("CourseHandler")
 # CONFIGURATIONS / SET UP
 REPO_PATH = Path(__file__).resolve().parent
 # Extract download path for Data Sets
-DATA_SETS_PATH = REPO_PATH / "data-sets"
+
 WORK_PATH = REPO_PATH / "work"
 # Set chunk size for downloading (avoid unnecessary loading to memory)
 CHUNK_SIZE = 5242880  # 5 MB in bytes
@@ -36,8 +37,6 @@ def path_already_exists(dir_path):
 
 
 def download():
-    # Create data-sets folder if it does not yet exist
-    create_dir_if_not_exists(DATA_SETS_PATH)
 
     # PROCESSING DATA SETS DEFINED IN CONFIGURATION FILE
     config = configparser.ConfigParser()
@@ -53,6 +52,8 @@ def download():
         filename = data_set_config["filename"]
         has_readme = data_set_config.get("has_readme") == "True"
         destination_path = data_set_config.get("destination_path")
+        file_type = data_set_config.get("type")
+
         if not has_readme:
             readme_location = data_set_config["readme_location"]
             license_info = data_set_config["license_info"]
@@ -62,11 +63,11 @@ def download():
 
         # Set download locations
         if destination_path:
-            destination_path = DATA_SETS_PATH / destination_path
+            destination_path = REPO_PATH / destination_path
             destination_downloadpath = destination_path
             destination_filepath = destination_path / filename
         else:
-            destination_path = DATA_SETS_PATH
+            destination_path = REPO_PATH
             destination_filepath = PurePath(destination_path / filename)
             destination_downloadpath = str(destination_filepath).replace(
                 destination_filepath.suffix, ""
@@ -76,50 +77,53 @@ def download():
         destination_filepath = str(destination_filepath)
 
         # Check if data is already downloaded
-        if path_already_exists(destination_downloadpath):
+        if not path_already_exists(destination_downloadpath):
             logger.info('Skipping "%s"', filename)
-            continue
+            # Create Directory
+            create_dir_if_not_exists(destination_path)
 
-        # Create Directory
-        create_dir_if_not_exists(destination_path)
+            # Create README.MD file (if needed)
+            if readme_md:
+                readme_loc = str(destination_path / "README.md")
+                with open(readme_loc, "wb") as readme:
+                    logger.info(" - Creating README.md file")
+                    readme.write(readme_md.encode("UTF-8"))
 
-        # Create README.MD file (if needed)
-        if readme_md:
-            readme_loc = str(destination_path / "README.md")
-            with open(readme_loc, "wb") as readme:
-                logger.info(" - Creating README.md file")
-                readme.write(readme_md.encode("UTF-8"))
-
-        # Download the file in chunks
-        with requests.get(str(download_path), stream=True, verify=False) as r:
-            r.raise_for_status()
-            logger.info(
-                'Downloading "%s" to "%s", this may take a few minutes',
-                filename,
-                destination_path,
-            )
-            with open(destination_filepath, "wb") as f:
-                for chunk in r.iter_content(CHUNK_SIZE):
-                    if chunk:  # filter out keep-alive chunks
-                        f.write(chunk)
+            # Download the file in chunks
+            with requests.get(str(download_path), stream=True, verify=False) as r:
+                r.raise_for_status()
+                logger.info(
+                    'Downloading "%s" to "%s", this may take a few minutes',
+                    filename,
+                    destination_path,
+                )
+                with open(destination_filepath, "wb") as f:
+                    for chunk in r.iter_content(CHUNK_SIZE):
+                        if chunk:  # filter out keep-alive chunks
+                            f.write(chunk)
+                        logger.info(
+                            "%.0f MB downloaded...",
+                            os.path.getsize(destination_filepath) / 1024 / 1024,
+                        )
+            logger.info("Finished downloading %s", filename)
+        if Path(destination_filepath).exists():
+            if file_type == 'zip':
+                # Extract zipfile
+                with ZipFile(destination_filepath, "r") as downloaded_file:
                     logger.info(
-                        "%.0f MB downloaded...",
-                        os.path.getsize(destination_filepath) / 1024 / 1024,
+                        'Extracting "%s" to "%s"',
+                        downloaded_file.filename,
+                        destination_filepath,
                     )
-        logger.info("Finished downloading %s", filename)
+                    downloaded_file.extractall(destination_path)
+            elif file_type == 'tgz':
+                print(destination_filepath + str(Path(destination_filepath).exists()))
+                subprocess.check_call(
+                    ['tar', '-xzf', destination_filepath, '-C', destination_path], shell=True)
 
-        # Extract zipfile
-        with ZipFile(destination_filepath, "r") as downloaded_file:
-            logger.info(
-                'Extracting "%s" to "%s"',
-                downloaded_file.filename,
-                destination_filepath,
-            )
-            downloaded_file.extractall(destination_path)
-
-        # Remove zip file
-        logger.info('Removing zip-file "%s"', filename)
-        os.remove(destination_filepath)
+            # Remove zip file
+            logger.info('Removing zip-file "%s"', filename)
+            os.remove(destination_filepath)
 
 
 if __name__ == "__main__":
